fix(dashboard): smooth subscription actions and consolidate tests
