+++ pages/dashboard/style.css
@@ -324,6 +324,8 @@ a:hover { text-decoration: underline; }
   border-bottom: 1px solid var(--border);
   margin-bottom: 12px;
 }
+.panel-head > :first-child { min-width: 0; }
+.panel-head > :first-child > * { min-width: 0; }
 .panel-head h2 { font-size: 16px; font-weight: 700; margin: 0; }
 .panel-head h3 { font-size: 14px; font-weight: 700; margin: 0; }
 
@@ -344,7 +346,6 @@ a:hover { text-decoration: underline; }
   transition: background var(--transition), border-color var(--transition), color var(--transition), transform var(--transition);
   line-height: 1.3;
 }
-.btn:active:not(:disabled) { transform: scale(0.97); }
 .btn:hover:not(:disabled) { background: var(--surface-hover); border-color: var(--blue-border); }
 .btn:active:not(:disabled) { transform: translateY(1px); }
 .btn:disabled { opacity: 0.5; cursor: not-allowed; }
@@ -386,9 +387,51 @@ a:hover { text-decoration: underline; }
   cursor: pointer;
   display: grid;
   place-items: center;
-  transition: all var(--transition);
+  transition: background var(--transition), border-color var(--transition), color var(--transition), transform var(--transition);
 }
 .btn-icon:hover:not(:disabled) { background: var(--blue-soft); color: var(--blue); }
+
+.input-action-row { flex-wrap: wrap; align-items: center; }
+.input-action-row > input,
+.input-action-row > textarea,
+.input-action-row > .dd {
+  flex: 1 1 220px;
+  min-width: 0;
+  width: auto !important;
+  max-width: 100%;
+}
+.input-action-row > .btn { flex: 0 0 auto; }
+
+body.ui-busy .view { pointer-events: none; cursor: wait; }
+body.ui-busy .view .btn,
+body.ui-busy .view .btn-icon { opacity: 0.6; cursor: wait; }
+body.ui-busy .view input,
+body.ui-busy .view textarea { cursor: wait; }
+
+.action-status {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin: 0 20px;
+  padding: 8px 12px;
+  border: 1px solid var(--blue-border);
+  border-radius: var(--radius-sm);
+  background: var(--blue-soft);
+  color: var(--blue);
+  font-size: 13px;
+  font-weight: 600;
+}
+.action-status::before {
+  content: "";
+  width: 13px;
+  height: 13px;
+  flex: 0 0 auto;
+  border: 2px solid currentColor;
+  border-right-color: transparent;
+  border-radius: 50%;
+  animation: status-spin 0.8s linear infinite;
+}
+@keyframes status-spin { to { transform: rotate(360deg); } }
 
 /* Form fields */
 .field { display: flex; flex-direction: column; gap: 5px; font-size: 13px; }
@@ -607,6 +650,7 @@ select { cursor: pointer; }
   gap: 16px;
   align-items: start;
 }
+.groups-layout > * { min-width: 0; }
 
 .group-list {
   display: flex;
