+++ tests/test_media_quality.py
@@ -1,14 +1,109 @@
-"""media_quality 三档选档测试。
-
-high=候选最高分辨率、low=候选最低、medium=候选中位数档（偶数取偏低档）。
-三档按候选内排序选档，不依赖宽高→p 换算，绕过横竖屏判定。
-"""
+"""media_quality contracts across HTML, status, and xdown media routes."""
 
 from __future__ import annotations
 
+import pytest
+
+from media_support.html_backend.parser import prefer_pbs_quality
 from media_support.service import MediaService
+from media_support.status_resolve import _apply_image_quality
 from media_support.xdown import XdownMediaCandidate
 from shared import TweetItem
+from shared.utils import TweetMedia
+
+PBS = "https://pbs.twimg.com/media/Gx1b48AJecgip5fZ.jpg"
+
+
+def test_high_uses_orig():
+    assert "name=orig" in prefer_pbs_quality(PBS, "high")
+
+
+def test_medium_uses_large():
+    assert "name=large" in prefer_pbs_quality(PBS, "medium")
+
+
+def test_low_uses_small():
+    assert "name=small" in prefer_pbs_quality(PBS, "low")
+
+
+def test_default_quality_is_high():
+    assert "name=orig" in prefer_pbs_quality(PBS)
+
+
+def test_rewrites_existing_name_param():
+    url = "https://pbs.twimg.com/media/Gx1.jpg?name=small&format=jpg"
+    assert prefer_pbs_quality(url, "high") == (
+        "https://pbs.twimg.com/media/Gx1.jpg?name=orig&format=jpg"
+    )
+
+
+def test_non_pbs_url_unchanged():
+    url = "https://video.twimg.com/amplify_video/x/vid/avc1/720x1280/y.mp4?tag=29"
+    assert prefer_pbs_quality(url, "high") == url
+
+
+def test_profile_images_not_rewritten():
+    url = "https://pbs.twimg.com/profile_images/123/x.jpg"
+    assert prefer_pbs_quality(url, "high") == url
+
+
+def test_unknown_quality_falls_back_to_orig():
+    assert "name=orig" in prefer_pbs_quality(PBS, "garbage")
+
+
+def _tweet(*media: TweetMedia) -> TweetItem:
+    tweet = TweetItem(
+        text="hi",
+        link="https://x.com/nasa/status/1",
+        published="",
+    )
+    tweet.media = list(media)
+    return tweet
+
+
+ORIG = "https://pbs.twimg.com/media/Fabc123.jpg?name=orig"
+
+
+@pytest.mark.parametrize(
+    ("quality", "expected"),
+    [
+        ("high", "name=orig"),
+        ("medium", "name=large"),
+        ("low", "name=small"),
+    ],
+)
+def test_image_quality_tier_applied(quality, expected):
+    tweet = _tweet(TweetMedia(kind="image", url=ORIG))
+    _apply_image_quality(tweet, quality)
+    assert expected in tweet.media[0].url
+
+
+def test_unknown_quality_falls_back_to_high():
+    tweet = _tweet(TweetMedia(kind="image", url=ORIG))
+    _apply_image_quality(tweet, "ultra")
+    assert "name=orig" in tweet.media[0].url
+
+
+def test_image_without_name_param_gets_one():
+    tweet = _tweet(TweetMedia(kind="image", url="https://pbs.twimg.com/media/Fabc.jpg"))
+    _apply_image_quality(tweet, "low")
+    assert tweet.media[0].url.endswith("?name=small")
+
+
+def test_video_url_untouched():
+    # 视频档位由 xdown 的多分辨率候选决定，pbs 改写只针对图片。
+    video = "https://video.twimg.com/amplify_video/1/vid/avc1/720x1280/x.mp4?tag=29"
+    tweet = _tweet(TweetMedia(kind="video", url=video))
+    _apply_image_quality(tweet, "low")
+    assert tweet.media[0].url == video
+
+
+def test_non_pbs_image_untouched():
+    url = "https://example.com/pic/media/abc.jpg"
+    tweet = _tweet(TweetMedia(kind="image", url=url))
+    _apply_image_quality(tweet, "low")
+    assert tweet.media[0].url == url
+
 
 TWEET = TweetItem(text="", link="https://x.com/u/status/1", published="")
